docs: Some basica docs for each package.
I think I want to have most of the docs in the main readme but if you're landing on the package from npm then it's never seen. So maybe we have the same docs for all 3? I'm trying seperating it out for now, but there is never a time when you don't want all the info at once imho.